refactor: split OpenTelemetryExtension
