+++ src/DependencyInjection/ExtensionLoader/ExtensionLoaderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\DependencyInjection\ExtensionLoader;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+interface ExtensionLoaderInterface
+{
+    public function load(array $config, ContainerBuilder $container): void;
+}
